Refactor of ECKey signature related methods: signatureToKey / recoverFromSignature / verify
