+++ rskj-core/src/main/java/co/rsk/RskContext.java
@@ -90,6 +90,7 @@ import org.ethereum.core.genesis.BlockChainLoader;
 import org.ethereum.core.genesis.GenesisLoader;
 import org.ethereum.core.genesis.GenesisLoaderImpl;
 import org.ethereum.crypto.ECKey;
+import org.ethereum.crypto.signature.SignatureService;
 import org.ethereum.datasource.DataSourceWithCache;
 import org.ethereum.datasource.KeyValueDataSource;
 import org.ethereum.datasource.LevelDbDataSource;
@@ -287,7 +288,6 @@ public class RskContext implements NodeBootstrapper {
             if (rskSystemProperties.importEnabled()) {
                 getBootstrapImporter().importData();
             }
-
             nodeRunner = buildNodeRunner();
         }
 
@@ -848,6 +848,7 @@ public class RskContext implements NodeBootstrapper {
                     getRepositoryLocator()
             ));
         }
+        SignatureService.initialize(getRskSystemProperties());
         return Collections.unmodifiableList(internalServices);
     }
