+++ rskj-core/src/main/java/co/rsk/net/discovery/message/PeerDiscoveryMessage.java
@@ -20,14 +20,16 @@ package co.rsk.net.discovery.message;
 
 import co.rsk.net.NodeID;
 import org.apache.commons.lang3.builder.ToStringBuilder;
+import org.bouncycastle.util.BigIntegers;
+import org.bouncycastle.util.encoders.Hex;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.crypto.HashUtil;
+import org.ethereum.crypto.signature.ECDSASignature;
+import org.ethereum.crypto.signature.SignatureService;
 import org.ethereum.util.ByteUtil;
 import org.ethereum.util.RLPElement;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.bouncycastle.util.BigIntegers;
-import org.bouncycastle.util.encoders.Hex;
 
 import java.security.SignatureException;
 import java.util.Optional;
@@ -65,7 +67,7 @@ public abstract class PeerDiscoveryMessage {
         byte[] forSig = HashUtil.keccak256(payload);
 
         /* [2] Crate signature*/
-        ECKey.ECDSASignature ecdsaSignature = privKey.sign(forSig);
+        ECDSASignature ecdsaSignature = privKey.sign(forSig);
 
         ecdsaSignature.v -= 27;
 
@@ -108,7 +110,7 @@ public abstract class PeerDiscoveryMessage {
 
         ECKey outKey = null;
         try {
-            outKey = ECKey.signatureToKey(msgHash, ECKey.ECDSASignature.fromComponents(r, s, v));
+            outKey = SignatureService.getInstance().signatureToKey(msgHash, ECDSASignature.fromComponents(r, s, v));
         } catch (SignatureException e) {
             logger.error("Error generating key from message", e);
         }
