+++ rskj-core/src/main/java/co/rsk/validators/ProofOfWorkRule.java
@@ -35,6 +35,8 @@ import org.ethereum.config.blockchain.upgrades.ConsensusRule;
 import org.ethereum.core.Block;
 import org.ethereum.core.BlockHeader;
 import org.ethereum.crypto.ECKey;
+import org.ethereum.crypto.signature.ECDSASignature;
+import org.ethereum.crypto.signature.SignatureService;
 import org.ethereum.util.RLP;
 import org.ethereum.util.RLPElement;
 import org.ethereum.util.RLPList;
@@ -260,7 +262,7 @@ public class ProofOfWorkRule implements BlockHeaderValidationRule, BlockValidati
             return false;
         }
 
-        ECKey.ECDSASignature signature = ECKey.ECDSASignature.fromComponents(r, s, v[0]);
+        ECDSASignature signature = ECDSASignature.fromComponents(r, s, v[0]);
 
         if (!Arrays.equals(r, signature.r.toByteArray())) {
             return false;
@@ -278,7 +280,7 @@ public class ProofOfWorkRule implements BlockHeaderValidationRule, BlockValidati
             return false;
         }
 
-        ECKey pub = ECKey.recoverFromSignature(signature.v - 27, signature, header.getHashForMergedMining(), false);
+        ECKey pub = SignatureService.getInstance().recoverFromSignature(signature.v - 27, signature, header.getHashForMergedMining(), false);
 
         return pub.getPubKeyPoint().equals(fallbackMiningPubKey.getPubKeyPoint());
     }
