+++ rskj-core/src/main/java/org/ethereum/core/Transaction.java
@@ -33,9 +33,10 @@ import org.bouncycastle.util.encoders.Hex;
 import org.ethereum.config.Constants;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.crypto.ECKey;
-import org.ethereum.crypto.ECKey.ECDSASignature;
+import org.ethereum.crypto.signature.ECDSASignature;
 import org.ethereum.crypto.ECKey.MissingPrivateKeyException;
 import org.ethereum.crypto.HashUtil;
+import org.ethereum.crypto.signature.SignatureService;
 import org.ethereum.util.ByteUtil;
 import org.ethereum.util.RLP;
 import org.ethereum.util.RLPList;
@@ -371,7 +372,7 @@ public class Transaction {
         Metric metric = profiler.start(Profiler.PROFILING_TYPE.KEY_RECOV_FROM_SIG);
         byte[] raw = getRawHash().getBytes();
         //We clear the 4th bit, the compress bit, in case a signature is using compress in true
-        ECKey key = ECKey.recoverFromSignature((signature.v - 27) & ~4, signature, raw, true);
+        ECKey key = SignatureService.getInstance().recoverFromSignature((signature.v - 27) & ~4, signature, raw, true);
         profiler.stop(metric);
         return key;
     }
@@ -384,7 +385,7 @@ public class Transaction {
 
         Metric metric = profiler.start(Profiler.PROFILING_TYPE.KEY_RECOV_FROM_SIG);
         try {
-            ECKey key = ECKey.signatureToKey(getRawHash().getBytes(), getSignature());
+            ECKey key = SignatureService.getInstance().signatureToKey(getRawHash().getBytes(), getSignature());
             sender = new RskAddress(key.getAddress());
         } catch (SignatureException e) {
             logger.error(e.getMessage(), e);
