+++ rskj-core/src/main/java/org/ethereum/crypto/signature/SignatureServiceBC.java
@@ -1,0 +1,171 @@
+/*
+ * This file is part of RskJ
+ * Copyright (C) 2020 RSK Labs Ltd.
+ * (derived from ethereumJ library, Copyright (c) 2016 <ether.camp>)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.ethereum.crypto.signature;
+
+import org.bouncycastle.asn1.sec.SECNamedCurves;
+import org.bouncycastle.asn1.x9.X9ECParameters;
+import org.bouncycastle.asn1.x9.X9IntegerConverter;
+import org.bouncycastle.crypto.params.ECDomainParameters;
+import org.bouncycastle.crypto.params.ECPublicKeyParameters;
+import org.bouncycastle.crypto.signers.ECDSASigner;
+import org.bouncycastle.math.ec.ECAlgorithms;
+import org.bouncycastle.math.ec.ECCurve;
+import org.bouncycastle.math.ec.ECPoint;
+import org.ethereum.crypto.ECKey;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.annotation.Nullable;
+import java.math.BigInteger;
+import java.security.SecureRandom;
+import java.security.SignatureException;
+
+/**
+ * Implementation of SignatureService with Bouncy Castle.
+ */
+class SignatureServiceBC extends SignatureService {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(SignatureServiceBC.class);
+    /**
+     * The parameters of the secp256k1 curve that Ethereum uses.
+     */
+    public static final ECDomainParameters CURVE;
+    public static final BigInteger HALF_CURVE_ORDER;
+    private static final SecureRandom secureRandom;
+
+    static {
+        // All clients must agree on the curve to use by agreement. Ethereum uses secp256k1.
+        X9ECParameters params = SECNamedCurves.getByName("secp256k1");
+        CURVE = new ECDomainParameters(params.getCurve(), params.getG(), params.getN(), params.getH());
+        HALF_CURVE_ORDER = params.getN().shiftRight(1);
+        secureRandom = new SecureRandom();
+    }
+
+    /**
+     * Part of the Singleton Signature service.
+     * {@link SignatureService#getInstance()}
+     */
+    SignatureServiceBC(){}
+
+    @Override
+    public ECKey signatureToKey(byte[] messageHash, ECDSASignature signature) throws SignatureException {
+        int header = signature.v & 0xFF;
+        // The header byte: 0x1B = first key with even y, 0x1C = first key with odd y,
+        //                  0x1D = second key with even y, 0x1E = second key with odd y
+        if (header < 27 || header > 34) {
+            throw new SignatureException("Header byte out of range: " + header);
+        }
+
+        boolean compressed = false;
+        if (header >= 31) {
+            compressed = true;
+            header -= 4;
+        }
+        int recId = header - 27;
+        ECKey key = this.recoverFromSignature(recId, signature, messageHash, compressed);
+        if (key == null) {
+            throw new SignatureException("Could not recover public key from signature");
+        }
+        return key;
+    }
+
+    @Nullable
+    @Override
+    public ECKey recoverFromSignature(int recId, ECDSASignature sig, byte[] messageHash, boolean compressed) {
+        check(recId >= 0, "recId must be positive");
+        check(sig.r.signum() >= 0, "r must be positive");
+        check(sig.s.signum() >= 0, "s must be positive");
+        check(messageHash != null, "messageHash must not be null");
+        // 1.0 For j from 0 to h   (h == recId here and the loop is outside this function)
+        //   1.1 Let x = r + jn
+        BigInteger n = CURVE.getN();  // Curve order.
+        BigInteger i = BigInteger.valueOf((long) recId / 2);
+        BigInteger x = sig.r.add(i.multiply(n));
+        //   1.2. Convert the integer x to an octet string X of length mlen using the conversion routine
+        //        specified in Section 2.3.7, where mlen = ⌈(log2 p)/8⌉ or mlen = ⌈m/8⌉.
+        //   1.3. Convert the octet string (16 set binary digits)||X to an elliptic curve point R using the
+        //        conversion routine specified in Section 2.3.4. If this conversion routine outputs “invalid”, then
+        //        do another iteration of Step 1.
+        //
+        // More concisely, what these points mean is to use X as a compressed public key.
+        ECCurve.Fp curve = (ECCurve.Fp) CURVE.getCurve();
+        BigInteger prime = curve.getQ();  // Bouncy Castle is not consistent about the letter it uses for the prime.
+        if (x.compareTo(prime) >= 0) {
+            // Cannot have point co-ordinates larger than this as everything takes place modulo Q.
+            return null;
+        }
+        // Compressed keys require you to know an extra bit of data about the y-coord as there are two possibilities.
+        // So it's encoded in the recId.
+        ECPoint r = decompressKey(x, (recId & 1) == 1);
+        //   1.4. If nR != point at infinity, then do another iteration of Step 1 (callers responsibility).
+        if (!r.multiply(n).isInfinity()) {
+            return null;
+        }
+        //   1.5. Compute e from M using Steps 2 and 3 of ECDSA signature verification.
+        BigInteger e = new BigInteger(1, messageHash);
+        //   1.6. For k from 1 to 2 do the following.   (loop is outside this function via iterating recId)
+        //   1.6.1. Compute a candidate public key as:
+        //               Q = mi(r) * (sR - eG)
+        //
+        // Where mi(x) is the modular multiplicative inverse. We transform this into the following:
+        //               Q = (mi(r) * s ** R) + (mi(r) * -e ** G)
+        // Where -e is the modular additive inverse of e, that is z such that z + e = 0 (mod n). In the above equation
+        // ** is point multiplication and + is point addition (the EC group operator).
+        //
+        // We can find the additive inverse by subtracting e from zero then taking the mod. For example the additive
+        // inverse of 3 modulo 11 is 8 because 3 + 8 mod 11 = 0, and -3 mod 11 = 8.
+        BigInteger eInv = BigInteger.ZERO.subtract(e).mod(n);
+        BigInteger rInv = sig.r.modInverse(n);
+        BigInteger srInv = rInv.multiply(sig.s).mod(n);
+        BigInteger eInvrInv = rInv.multiply(eInv).mod(n);
+        ECPoint.Fp q = (ECPoint.Fp) ECAlgorithms.sumOfTwoMultiplies(CURVE.getG(), eInvrInv, r, srInv);
+        return ECKey.fromPublicOnly(q.getEncoded(compressed));
+    }
+
+    @Override
+    public boolean verify(byte[] data, ECDSASignature signature, byte[] pub) {
+        ECDSASigner signer = new ECDSASigner();
+        ECPublicKeyParameters params = new ECPublicKeyParameters(CURVE.getCurve().decodePoint(pub), CURVE);
+        signer.init(false, params);
+        try {
+            return signer.verifySignature(data, signature.r, signature.s);
+        } catch (NullPointerException npe) {
+            // Bouncy Castle contains a bug that can cause NPEs given specially crafted signatures.
+            // Those signatures are inherently invalid/attack sigs so we just fail them here rather than crash the thread.
+            LOGGER.error("Caught NPE inside bouncy castle", npe);
+            return false;
+        }
+    }
+
+    /**
+     * Decompress a compressed public key (x co-ord and low-bit of y-coord).
+     *
+     * @param xBN -
+     * @param yBit -
+     * @return -
+     */
+    private ECPoint decompressKey(BigInteger xBN, boolean yBit) {
+        X9IntegerConverter x9 = new X9IntegerConverter();
+        byte[] compEnc = x9.integerToBytes(xBN, 1 + x9.getByteLength(CURVE.getCurve()));
+        compEnc[0] = (byte) (yBit ? 0x03 : 0x02);
+        return CURVE.getCurve().decodePoint(compEnc);
+    }
+
+}
