+++ rskj-core/src/main/java/org/ethereum/net/rlpx/AuthInitiateMessage.java
@@ -22,6 +22,7 @@ package org.ethereum.net.rlpx;
 import org.ethereum.crypto.ECKey;
 import org.bouncycastle.math.ec.ECPoint;
 import org.bouncycastle.util.encoders.Hex;
+import org.ethereum.crypto.signature.ECDSASignature;
 
 import static org.ethereum.util.ByteUtil.merge;
 import static org.bouncycastle.util.BigIntegers.asUnsignedByteArray;
@@ -32,7 +33,7 @@ import static org.bouncycastle.util.BigIntegers.asUnsignedByteArray;
  * Created by devrandom on 2015-04-07.
  */
 public class AuthInitiateMessage {
-    ECKey.ECDSASignature signature; // 65 bytes
+    ECDSASignature signature; // 65 bytes
     byte[] ephemeralPublicHash; // 32 bytes
     ECPoint publicKey; // 64 bytes - uncompressed and no type byte
     byte[] nonce; // 32 bytes
@@ -56,7 +57,7 @@ public class AuthInitiateMessage {
         offset += 32;
         int v = wire[offset] + 27;
         offset += 1;
-        message.signature = ECKey.ECDSASignature.fromComponents(r, s, (byte)v);
+        message.signature = ECDSASignature.fromComponents(r, s, (byte)v);
         message.ephemeralPublicHash = new byte[32];
         System.arraycopy(wire, offset, message.ephemeralPublicHash, 0, 32);
         offset += 32;
