+++ rskj-core/src/main/java/org/ethereum/rpc/dto/TransactionResultDTO.java
@@ -22,7 +22,7 @@ import co.rsk.core.Coin;
 import co.rsk.remasc.RemascTransaction;
 import org.ethereum.core.Block;
 import org.ethereum.core.Transaction;
-import org.ethereum.crypto.ECKey;
+import org.ethereum.crypto.signature.ECDSASignature;
 import org.ethereum.rpc.TypeConverter;
 import org.ethereum.util.ByteUtil;
 
@@ -78,7 +78,7 @@ public class TransactionResultDTO {
         input = TypeConverter.toUnformattedJsonHex(tx.getData());
 
         if (!(tx instanceof RemascTransaction)) {
-            ECKey.ECDSASignature signature = tx.getSignature();
+            ECDSASignature signature = tx.getSignature();
             v = String.format("0x%02x", signature.v);
             r = TypeConverter.toQuantityJsonHex(signature.r);
             s = TypeConverter.toQuantityJsonHex(signature.s);
