+++ rskj-core/src/main/java/org/ethereum/vm/PrecompiledContracts.java
@@ -39,6 +39,8 @@ import org.ethereum.core.Repository;
 import org.ethereum.core.Transaction;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.crypto.HashUtil;
+import org.ethereum.crypto.signature.ECDSASignature;
+import org.ethereum.crypto.signature.SignatureService;
 import org.ethereum.db.BlockStore;
 import org.ethereum.db.ReceiptStore;
 import org.ethereum.util.BIUtil;
@@ -311,9 +313,9 @@ public class PrecompiledContracts {
                 System.arraycopy(data, 96, s, 0, sLength);
 
                 if (isValid(r, s, v)) {
-                    ECKey.ECDSASignature signature = ECKey.ECDSASignature.fromComponents(r, s, v[31]);
+                    ECDSASignature signature = ECDSASignature.fromComponents(r, s, v[31]);
 
-                    ECKey key = ECKey.signatureToKey(h, signature);
+                    ECKey key = SignatureService.getInstance().signatureToKey(h, signature);
                     out = DataWord.valueOf(key.getAddress());
                 }
             } catch (Exception any) {
@@ -332,7 +334,7 @@ public class PrecompiledContracts {
             BigInteger r = new BigInteger(1, rBytes);
             BigInteger s = new BigInteger(1, sBytes);
 
-            return ECKey.ECDSASignature.validateComponents(r, s, v);
+            return ECDSASignature.validateComponents(r, s, v);
         }
     }
