+++ rskj-core/src/main/resources/expected.conf
@@ -238,6 +238,10 @@ cache = {
   }
 }
 
+crypto = {
+    library = <bc-or-native>
+}
+
 GitHubTests.VMTest.loadLocal = <bool>
 messages.recorder = {
     enabled = <bool>
