+++ rskj-core/src/main/resources/reference.conf
@@ -381,3 +381,11 @@ cache {
     max-elements: 100000
   }
 }
+
+crypto {
+    #Possibility to config crypto library implementation to use
+    # Possible values:
+    # - 'bc' (Bouncy Castle)
+    # - 'native' (Native C lib) <- not yet implemented
+    library: 'bc'
+}
