+++ rskj-core/src/test/java/org/ethereum/crypto/ECKeyTest.java
@@ -26,7 +26,7 @@ import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.common.util.concurrent.MoreExecutors;
 import org.ethereum.core.ImmutableTransaction;
 import org.ethereum.core.Transaction;
-import org.ethereum.crypto.ECKey.ECDSASignature;
+import org.ethereum.crypto.signature.ECDSASignature;
 import org.ethereum.util.ByteUtil;
 import org.junit.Assert;
 import org.junit.Ignore;
@@ -142,80 +142,29 @@ public class ECKeyTest {
     }
 
     @Test
-    public void testVerifySignature1() {
-        ECKey key = ECKey.fromPublicOnly(pubKey);
-        BigInteger r = new BigInteger("28157690258821599598544026901946453245423343069728565040002908283498585537001");
-        BigInteger s = new BigInteger("30212485197630673222315826773656074299979444367665131281281249560925428307087");
-        ECDSASignature sig = ECDSASignature.fromComponents(r.toByteArray(), s.toByteArray(), (byte) 28);
-        key.verify(HashUtil.keccak256(exampleMessage.getBytes()), sig);
-    }
-
-    @Test
-    public void testVerifySignature2() {
-        BigInteger r = new BigInteger("c52c114d4f5a3ba904a9b3036e5e118fe0dbb987fe3955da20f2cd8f6c21ab9c", 16);
-        BigInteger s = new BigInteger("6ba4c2874299a55ad947dbc98a25ee895aabf6b625c26c435e84bfd70edf2f69", 16);
-        ECDSASignature sig = ECDSASignature.fromComponents(r.toByteArray(), s.toByteArray(), (byte) 0x1b);
-        byte[] rawtx = Hex.decode("f82804881bc16d674ec8000094cd2a3d9f938e13cd947ec05abc7fe734df8dd8268609184e72a0006480");
-        try {
-            ECKey key = ECKey.signatureToKey(HashUtil.keccak256(rawtx), sig);
-            System.out.println("Signature public key\t: " + Hex.toHexString(key.getPubKey()));
-            System.out.println("Sender is\t\t: " + Hex.toHexString(key.getAddress()));
-            assertEquals("cd2a3d9f938e13cd947ec05abc7fe734df8dd826", Hex.toHexString(key.getAddress()));
-            key.verify(HashUtil.keccak256(rawtx), sig);
-        } catch (SignatureException e) {
-            fail();
-        }
-    }
-
-    @Test
-    @Ignore("The TX sender is not 20-byte long")
-    public void testVerifySignature3() throws SignatureException {
-
-        byte[] rawtx = Hex.decode("f86e80893635c9adc5dea000008609184e72a00082109f9479b08ad8787060333663d19704909ee7b1903e58801ba0899b92d0c76cbf18df24394996beef19c050baa9823b4a9828cd9b260c97112ea0c9e62eb4cf0a9d95ca35c8830afac567619d6b3ebee841a3c8be61d35acd8049");
-
-        Transaction tx = new ImmutableTransaction(rawtx);
-        ECKey key = ECKey.signatureToKey(HashUtil.keccak256(rawtx), tx.getSignature());
-
-        System.out.println("Signature public key\t: " + Hex.toHexString(key.getPubKey()));
-        System.out.println("Sender is\t\t: " + Hex.toHexString(key.getAddress()));
-
-        //  sender: CD2A3D9F938E13CD947EC05ABC7FE734DF8DD826
-        // todo: add test assertion when the sign/verify part actually works.
-    }
-
-
-    @Test
     public void testSValue() throws Exception {
         // Check that we never generate an S value that is larger than half the curve order. This avoids a malleability
         // issue that can allow someone to change a transaction [hash] without invalidating the signature.
         final int ITERATIONS = 10;
         ListeningExecutorService executor = MoreExecutors.listeningDecorator(Executors.newFixedThreadPool(ITERATIONS));
-        List<ListenableFuture<ECKey.ECDSASignature>> sigFutures = Lists.newArrayList();
+        List<ListenableFuture<ECDSASignature>> sigFutures = Lists.newArrayList();
         final ECKey key = new ECKey();
         for (byte i = 0; i < ITERATIONS; i++) {
             final byte[] hash = HashUtil.keccak256(new byte[]{i});
             sigFutures.add(executor.submit(new Callable<ECDSASignature>() {
                 @Override
-                public ECKey.ECDSASignature call() throws Exception {
+                public ECDSASignature call() throws Exception {
                     return key.doSign(hash);
                 }
             }));
         }
-        List<ECKey.ECDSASignature> sigs = Futures.allAsList(sigFutures).get();
-        for (ECKey.ECDSASignature signature : sigs) {
+        List<ECDSASignature> sigs = Futures.allAsList(sigFutures).get();
+        for (ECDSASignature signature : sigs) {
             assertTrue(signature.s.compareTo(ECKey.HALF_CURVE_ORDER) <= 0);
         }
-        final ECKey.ECDSASignature duplicate = new ECKey.ECDSASignature(sigs.get(0).r, sigs.get(0).s);
+        final ECDSASignature duplicate = new ECDSASignature(sigs.get(0).r, sigs.get(0).s);
         assertEquals(sigs.get(0), duplicate);
         assertEquals(sigs.get(0).hashCode(), duplicate.hashCode());
-    }
-
-    @Test
-    public void testSignVerify() {
-        ECKey key = ECKey.fromPrivate(privateKey);
-        String message = "This is an example of a signed message.";
-        ECDSASignature output = key.doSign(message.getBytes());
-        assertTrue(key.verify(message.getBytes(), output));
     }
 
     @Test
@@ -264,36 +213,6 @@ public class ECKeyTest {
     }
 
     @Test
-    public void keyRecovery() throws Exception {
-        ECKey key = new ECKey();
-        String message = "Hello World!";
-        byte[] hash = HashUtil.sha256(message.getBytes());
-        ECKey.ECDSASignature sig = key.doSign(hash);
-        key = ECKey.fromPublicOnly(key.getPubKeyPoint());
-        boolean found = false;
-        for (int i = 0; i < 4; i++) {
-            ECKey key2 = ECKey.recoverFromSignature(i, sig, hash, true);
-            checkNotNull(key2);
-            if (key.equals(key2)) {
-                found = true;
-                break;
-            }
-        }
-        assertTrue(found);
-    }
-
-    @Test
-    public void testSignedMessageToKey() throws SignatureException {
-        byte[] messageHash = HashUtil.keccak256(exampleMessage.getBytes());
-
-        ECKey key = ECKey.signatureToKey(messageHash, ECDSASignature.fromComponents(ByteUtil.bigIntegerToBytes(r, 32),
-                ByteUtil.bigIntegerToBytes(s, 32), v));
-
-        assertNotNull(key);
-        assertArrayEquals(pubKey, key.getPubKey());
-    }
-
-    @Test
     public void testGetPrivKeyBytes() {
         ECKey key = new ECKey();
         assertNotNull(key.getPrivKeyBytes());
@@ -311,11 +230,11 @@ public class ECKeyTest {
         assertTrue(key1.equals(key2));
     }
 
-
     @Test
     public void decryptAECSIC(){
         ECKey key = ECKey.fromPrivate(Hex.decode("abb51256c1324a1350598653f46aa3ad693ac3cf5d05f36eba3f495a1f51590f"));
         byte[] payload = key.decryptAES(Hex.decode("84a727bc81fa4b13947dc9728b88fd08"));
         System.out.println(Hex.toHexString(payload));
     }
+
 }
