+++ rskj-core/src/test/java/org/ethereum/crypto/signature/ECDSASignatureTest.java
@@ -1,6 +1,7 @@
 /*
  * This file is part of RskJ
  * Copyright (C) 2017 RSK Labs Ltd.
+ * (derived from ethereumJ library, Copyright (c) 2016 <ether.camp>)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -16,28 +17,23 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package co.rsk.crypto;
+package org.ethereum.crypto.signature;
 
-import org.ethereum.TestUtils;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.crypto.HashUtil;
 import org.junit.Assert;
 import org.junit.Test;
-import org.bouncycastle.util.encoders.Base64;
-import java.security.SignatureException;
-import static org.junit.Assert.*;
 
-public class ECKeyTest {
-    private String exampleMessage = new String("This is an example of a signed message.");
+public class ECDSASignatureTest {
 
     @Test
     public void fromComponentsWithRecoveryCalculation() {
         ECKey key = new ECKey();
         byte[] hash = HashUtil.randomHash();
-        ECKey.ECDSASignature signature = key.sign(hash);
+        ECDSASignature signature = key.sign(hash);
 
         // With uncompressed public key
-        ECKey.ECDSASignature signatureWithCalculatedV = ECKey.ECDSASignature.fromComponentsWithRecoveryCalculation(
+        ECDSASignature signatureWithCalculatedV = fromComponentsWithRecoveryCalculation(
                 signature.r.toByteArray(),
                 signature.s.toByteArray(),
                 hash,
@@ -49,7 +45,7 @@ public class ECKeyTest {
         Assert.assertEquals(signature.v, signatureWithCalculatedV.v);
 
         // With compressed public key
-        signatureWithCalculatedV = ECKey.ECDSASignature.fromComponentsWithRecoveryCalculation(
+        signatureWithCalculatedV = fromComponentsWithRecoveryCalculation(
                 signature.r.toByteArray(),
                 signature.s.toByteArray(),
                 hash,
@@ -59,5 +55,41 @@ public class ECKeyTest {
         Assert.assertEquals(signature.r, signatureWithCalculatedV.r);
         Assert.assertEquals(signature.s, signatureWithCalculatedV.s);
         Assert.assertEquals(signature.v, signatureWithCalculatedV.v);
+    }
+
+
+    /**
+     *
+     * @param r -
+     * @param s -
+     * @param hash - the hash used to compute this signature
+     * @param pub - public key bytes, used to calculate the recovery byte 'v'
+     * @return -
+     */
+    public static ECDSASignature fromComponentsWithRecoveryCalculation(byte[] r, byte[] s, byte[] hash, byte[] pub) {
+        byte v = calculateRecoveryByte(r, s, hash, pub);
+        return ECDSASignature.fromComponents(r, s, v);
+    }
+
+    private static byte calculateRecoveryByte(byte[] r, byte[] s, byte[] hash, byte[] pub) {
+        ECDSASignature sig = ECDSASignature.fromComponents(r, s);
+        ECKey pubKey = ECKey.fromPublicOnly(pub);
+
+        // TODO: same logic in ECKey.sign
+        // Now we have to work backwards to figure out the recId needed to recover the signature.
+        int recId = -1;
+        for (int i = 0; i < 4; i++) {
+            ECKey k = SignatureService.getInstance().recoverFromSignature(i, sig, hash, false);
+            if (k != null && k.equalsPub(pubKey)) {
+                recId = i;
+                break;
+            }
+        }
+
+        if (recId == -1) {
+            throw new RuntimeException("Could not construct a recoverable key. This should never happen.");
+        }
+
+        return (byte) (recId + 27);
     }
 }
