+++ rskj-core/src/test/java/org/ethereum/net/rlpx/EIP8HandshakeTest.java
@@ -20,6 +20,7 @@
 package org.ethereum.net.rlpx;
 
 import org.ethereum.crypto.ECKey;
+import org.ethereum.crypto.signature.ECDSASignature;
 import org.junit.Before;
 import org.junit.Test;
 import org.bouncycastle.crypto.InvalidCipherTextException;
@@ -40,7 +41,7 @@ public class EIP8HandshakeTest {
     private ECKey ephemeralKeyB = fromPrivate(decode("e238eb8e04fee6511ab04c6dd3c89ce097b11f25d584863ac2b6d5b35b1847e4"));
     private byte[] nonceA = decode("7e968bba13b6c50e2c4cd7f241cc0d64d1ac25c7f5952df231ac6a2bda8ee5d6");
     private byte[] nonceB = decode("559aead08264d5795d3909718cdd05abd49572e84fe55590eef31a88a08fdffd");
-    private ECKey.ECDSASignature signatureA = ECKey.ECDSASignature.fromComponents(
+    private ECDSASignature signatureA = ECDSASignature.fromComponents(
             decode("299ca6acfd35e3d72d8ba3d1e2b60b5561d5af5218eb5bc182045769eb422691"),
             decode("0a301acae3b369fffc4a4899d6b02531e89fd4fe36a2cf0d93607ba470b50f78"),
             (byte) 27
